# backend/swagger/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My API",
    "description": "스웨거 자동생성",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gallery": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "gallery_title": {
                    "example": "any"
                  },
                  "gallery_image_url": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/gallery/{gallery_id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "gallery_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "gallery_title": {
                    "example": "any"
                  },
                  "gallery_image_url": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "gallery_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/blog": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "blog_title": {
                    "example": "any"
                  },
                  "blog_writer": {
                    "example": "any"
                  },
                  "blog_content": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/blog/{blog_id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "blog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "blog_title": {
                    "example": "any"
                  },
                  "blog_writer": {
                    "example": "any"
                  },
                  "blog_content": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "blog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}